Enable autogeneration for default runtime metrics list in collectors tests according to Go version
