+++ prometheus/collectors/go_collector_go119_test.go
@@ -16,8 +16,6 @@
 
 package collectors
 
-import "sort"
-
 func withAllMetrics() []string {
 	return withBaseMetrics([]string{
 		"go_cgo_go_to_c_calls_calls_total",
@@ -110,19 +108,4 @@ func withSchedulerMetrics() []string {
 
 func withDebugMetrics() []string {
 	return withBaseMetrics([]string{})
-}
-
-var defaultRuntimeMetrics = []string{
-	"go_sched_gomaxprocs_threads",
-}
-
-func withDefaultRuntimeMetrics(metricNames []string, withoutGC, withoutSched bool) []string {
-	// If withoutSched is true, exclude "go_sched_gomaxprocs_threads".
-	if withoutSched {
-		return metricNames
-	}
-	metricNames = append(metricNames, defaultRuntimeMetrics...)
-	// sorting is required
-	sort.Strings(metricNames)
-	return metricNames
 }
