+++ prometheus/collectors/go_collector_latest_test.go
@@ -62,6 +62,25 @@ var memstatMetrics = []string{
 	"go_memstats_sys_bytes",
 }
 
+func withDefaultRuntimeMetrics(metricNames []string, withoutGC, withoutSched bool) []string {
+	if withoutGC && withoutSched {
+		// If both flags are true, return the metricNames as is.
+		return metricNames
+	} else if withoutGC && !withoutSched {
+		// If only withoutGC is true, include "go_sched_gomaxprocs_threads" only.
+		metricNames = append(metricNames, []string{"go_sched_gomaxprocs_threads"}...)
+	} else if withoutSched && !withoutGC {
+		// If only withoutSched is true, exclude "go_sched_gomaxprocs_threads".
+		metricNames = append(metricNames, []string{"go_gc_gogc_percent", "go_gc_gomemlimit_bytes"}...)
+	} else {
+		// If neither flag is true, use the default metrics.
+		metricNames = append(metricNames, defaultRuntimeMetrics...)
+	}
+	// sorting is required
+	sort.Strings(metricNames)
+	return metricNames
+}
+
 func TestGoCollectorMarshalling(t *testing.T) {
 	reg := prometheus.NewPedanticRegistry()
 	reg.MustRegister(NewGoCollector(
